+++ game/path/MultiShapes.java
@@ -1,7 +1,7 @@
 package game.path;
 
-import javafx.scene.paint.Color;
 import javafx.scene.Group;
+import javafx.scene.paint.Color;
 
 
 public class MultiShapes extends Obstacle {
@@ -14,38 +14,38 @@ public class MultiShapes extends Obstacle {
 	 version 5 :
 	 */
 
-	public MultiShapes(int x, int y,Color[] colors,int version){
-		super(x,y, colors,version);
-		obstacle = buildObstacle();
-	}
-	
-	public Group buildObstacle(){
-		Group multiShapes = new Group();
-		
-		int length = 100;
-		int tinyRadial = 40;
-		int mediumRadial = 70;
-		int internRadial = 85;
-		int bigRadial = 100;
-		int width = 15;
-		int nb_arc = 4;
-		int nbr_versions = 6;
-		
-		switch(version){
-		case 0 :
-			break;
-		case 1 :
-			break;
-		case 2 :
-			break;
-		case 3 :
-			break;
-		case 4 :
-			break;
-		case 5 :
-			break;
-		}
-		
-		return multiShapes;
-	}
+    public MultiShapes(double x, double y, Color[] colors, int version) {
+        super(x, y, colors, version);
+        obstacle = buildObstacle();
+    }
+
+    public Group buildObstacle() {
+        Group multiShapes = new Group();
+
+        double length = 100.0;
+        double tinyRadial = 40.0;
+        double mediumRadial = 70.0;
+        double internRadial = 85.0;
+        double bigRadial = 100.0;
+        double width = 15.0;
+        int nb_arc = 4;
+        int nbr_versions = 6;
+
+        switch (version) {
+            case 0:
+                break;
+            case 1:
+                break;
+            case 2:
+                break;
+            case 3:
+                break;
+            case 4:
+                break;
+            case 5:
+                break;
+        }
+
+        return multiShapes;
+    }
 }
